refactor: extract client-types, useTriggersState, usePairingState
Phase 2-3 of core debt refactoring:

- client-types.ts: 1,848 lines of types extracted from client.ts (6,062 → 4,392)
- useTriggersState.ts: 6 useState + 8 callbacks for trigger CRUD (215 lines)
- usePairingState.ts: 5 useState + submit handler for pairing UI (63 lines)

AppContext.tsx: 8,366 → 8,150 lines. Each hook follows the useLifecycleState
pattern — state + actions returned, destructured in AppProviderInner.
Context value shape unchanged — zero consumer changes.
